+++ src/evaluate.zig
@@ -173,6 +173,40 @@ pub fn spaceBonus(pos: position.Position) types.Value {
     return bonus;
 }
 
+pub fn outpostBonus(pos: position.Position, comptime color: types.Color) types.Value {
+    var count: types.Value = 0;
+    var knights: types.Bitboard = pos.bb_colors[color.index()] & pos.bb_pieces[types.PieceType.knight.index()];
+    const pawns: types.Bitboard = pos.bb_colors[color.index()] & pos.bb_pieces[types.PieceType.pawn.index()];
+    const pawns_them: types.Bitboard = pos.bb_colors[color.invert().index()] & pos.bb_pieces[types.PieceType.pawn.index()];
+
+    // Remove unprotected knights
+    var knight_diagonal_left: types.Bitboard = undefined;
+    var knight_diagonal_right: types.Bitboard = undefined;
+    if (color.isWhite()) {
+        knight_diagonal_right = (knights & ~types.mask_file[0]) >> (types.board_size - 1);
+        knight_diagonal_left = (knights & ~types.mask_file[types.board_size - 1]) >> (types.board_size + 1);
+        const to_keep: types.Bitboard = (knight_diagonal_right & pawns) << (types.board_size - 1) | (knight_diagonal_left & pawns) << (types.board_size + 1);
+        knights &= to_keep;
+    } else {
+        knight_diagonal_right = (knights & ~types.mask_file[types.board_size - 1]) << (types.board_size + 1);
+        knight_diagonal_left = (knights & ~types.mask_file[0]) << (types.board_size - 1);
+        const to_keep: types.Bitboard = (knight_diagonal_right & pawns) >> (types.board_size + 1) | (knight_diagonal_left & pawns) >> (types.board_size - 1);
+        knights &= to_keep;
+    }
+
+    // Count outposts
+    // Maybe this could be done without the while
+    while (knights != 0) {
+        const sq: types.Square = types.popLsb(&knights);
+        // Use passed pawn filter without the current file
+        const attackers: types.Bitboard = (tables.passed_pawn[color.index()][sq.index()] & ~types.mask_file[sq.file().index()]) & pawns_them;
+        if (attackers == 0)
+            count += 1;
+    }
+    return count;
+}
+
+// TODO: Add pawn structure hash
 pub fn evaluateTable(pos: position.Position) types.Value {
     var score: types.Value = pos.score_material_w - pos.score_material_b;
     const endgame: bool = pos.endgame(pos.state.turn);
@@ -200,6 +234,8 @@ pub fn evaluateTable(pos: position.Position) types.Value {
     score +|= mobilityBonus(pos, .white) - mobilityBonus(pos, .black);
 
     score +|= spaceBonus(pos);
+
+    score +|= variable.getValue("outpost") *| (outpostBonus(pos, .white) - outpostBonus(pos, .white));
 
     const bishops: types.Bitboard = pos.bb_pieces[types.PieceType.bishop.index()];
     const white_pair: types.Value = @intFromBool(@popCount(bb_white & bishops) >= 2);